requirements.txt
Flask==3.0.2
yt-dlp
google-generativeai==0.4.1
python-dotenv==1.0.1
